requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
rich>=13.0.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=1.0.0
